requirements.txt
google-adk[a2a]>=1.0.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
litellm>=1.53.5
pyyaml>=6.0
